backend/requirements.txt
fastapi
uvicorn
psutil
PySensors
python-dotenv